+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved equivalence checker that avoids checking similar branches more than once.
 - Improved simplification for arithmetic expressions
 - Construction of storage counterexamples based on the model returned by the SMT solver.
+- Static binaries for macos
 
 ## [0.50.2] - 2023-01-06